depreciate NCPManager & Merge into InteractionManager & GrimManager -> AntiCheatManager
